% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-correlation.R
\name{calc_correlation}
\alias{calc_correlation}
\title{Get the correlation.}
\usage{
calc_correlation(
  dataset,
  id,
  dataset_correlate = NULL,
  intcovar = NULL,
  use_qr = TRUE,
  backfill_NA = TRUE,
  full_annotations = FALSE
)
}
\arguments{
\item{dataset}{The dataset object}

\item{id}{The unique id in the dataset.}

\item{dataset_correlate}{The dataset to correlate against, dataset if INVALID
or NULL.}

\item{intcovar}{The interactive covariate.}

\item{use_qr}{`TRUE` to use qr decomposition.}

\item{backfill_NA}{`TRUE` if we need to impute, we backfill the NAs.}

\item{full_annotations}{`TRUE` to send back full annotations.}
}
\value{
A `tibble` with the correlation and annotations.
}
\description{
Perform a correlation analysis on an `id` for a `dataset`.  By default, the
correlation will be against the same dataset, but a different dataset can
be specified with `dataset_correlate`.
}
