% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-correlation-plot.R
\name{calc_correlation_plot}
\alias{calc_correlation_plot}
\title{Get the correlation data for plotting.}
\usage{
calc_correlation_plot(
  dataset,
  id,
  dataset_correlate,
  id_correlate,
  intcovar = NULL,
  use_qr = TRUE,
  backfill_NA = TRUE
)
}
\arguments{
\item{dataset}{The dataset object}

\item{id}{The unique id in the dataset.}

\item{dataset_correlate}{The dataset to correlate against, dataset if INVALID
or NULL.}

\item{id_correlate}{The identifier from the correlate dataset.}

\item{intcovar}{The interactive covariate.}

\item{use_qr}{`TRUE` to use qr decomposition.}

\item{backfill_NA}{`TRUE` if we need to impute, we backfill the NAs.}
}
\value{
A named `list` with keys of `data` and `datatypes`.
}
\description{
Perform a correlation analysis on an `id` for a `dataset`.  By default, the
correlation will be against the same dataset, but a different dataset can
be specified with `dataset_correlate`.  Return x and y values for plotting.
}
