% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-lod-peaks-for-annot.R
\name{calc_lod_peaks_for_annot}
\alias{calc_lod_peaks_for_annot}
\title{Get all the peaks for an id.}
\usage{
calc_lod_peaks_for_annot(
  dataset,
  id,
  threshold = 6,
  peakdrop = 2,
  thresholdX = 6,
  peakdropX = 2,
  n_cores = 0,
  calc_diff = FALSE
)
}
\arguments{
\item{dataset}{the dataset object}

\item{id}{the unique id in the dataset}

\item{threshold}{if set, qtl2::find_peaks is used}

\item{peakdrop}{if set, qtl2::find_peaks is used}

\item{thresholdX}{if set, qtl2::find_peaks is used}

\item{peakdropX}{if set, qtl2::find_peaks is used}

\item{n_cores}{number of cores to use (0=ALL)}

\item{calc_diff}{compute the difference between additive and covariate}
}
\value{
a tibble of the peaks.
}
\description{
Get all the peaks for an id.
}
