% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-lod-scores-by-covar.R
\name{calc_lod_scores_by_covar}
\alias{calc_lod_scores_by_covar}
\title{Perform the LOD scan for each "value" of the interactive covariate.}
\usage{
calc_lod_scores_by_covar(dataset, id, chrom, intcovar, cores = 0)
}
\arguments{
\item{dataset}{the dataset object}

\item{id}{the unique id in the dataset}

\item{chrom}{The chromosome.}

\item{intcovar}{the interactive covariate}

\item{cores}{number of cores to $se (0=ALL)}
}
\value{
A named list with each name a "sample value" and the element is a
  tibble with the following columns: id, chr, pos, lod.
}
\description{
Perform the LOD scan for each "value" of the interactive covariate.
}
