% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-mediation.R
\name{calc_mediation}
\alias{calc_mediation}
\title{Perform mediation.}
\usage{
calc_mediation(
  dataset,
  id,
  marker_id,
  dataset_mediate = NULL,
  full_annotations = FALSE
)
}
\arguments{
\item{dataset}{The dataset object.}

\item{id}{The unique id in the dataset.}

\item{marker_id}{The unique marker identifier.}

\item{dataset_mediate}{The dataset object to mediate against (defaults to
dataset)}

\item{full_annotations}{`TRUE` to send back full annotations.}
}
\value{
A `tibble` with the following columns depending on datatype
}
\description{
Perform mediation.
}
