% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-snp-assoc-mapping.R
\name{calc_snp_assoc_mapping}
\alias{calc_snp_assoc_mapping}
\title{Get the SNP association mapping from foundersnps database.}
\usage{
calc_snp_assoc_mapping(
  dataset,
  id,
  chrom,
  location,
  db_file,
  window_size = 5e+05,
  intcovar = NULL,
  cores = 0
)
}
\arguments{
\item{dataset}{The dataset object.}

\item{id}{The unique id in the dataset.}

\item{chrom}{The chromosome.}

\item{location}{Location on chromosome in base pairs.}

\item{db_file}{Full path to the sqlite database file.}

\item{window_size}{The size of the window to scan before and after location.}

\item{intcovar}{The interactive covariate.}

\item{cores}{Number of cores to use (0 = ALL).}
}
\value{
a `tibble` with the following columns:
      snp, chr, pos, alleles, sdp, ensembl_gene, csq, index, interval,
      on_map, lod
}
\description{
Get the SNP association mapping from foundersnps database.
}
