% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_datasets}
\alias{compare_datasets}
\title{Compare 2 datasets.}
\usage{
compare_datasets(dataset_a, dataset_b)
}
\arguments{
\item{dataset_a}{the dataset_id as a string identifier}

\item{dataset_b}{the dataset_id as a string identifier}
}
\description{
Compare 2 datasets.
}
