% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{\%<>\%}
\alias{\%<>\%}
\title{Compound assignment operator}
\usage{
lhs \%<>\% rhs
}
\arguments{
\item{lhs}{A value or the magrittr placeholder.}

\item{rhs}{A function call using the magrittr semantics.}
}
\value{
The result of calling \code{rhs(lhs)}.
}
\description{
See \code{magrittr::\link[magrittr:compound]{\%<>\%}} for details.
}
\keyword{internal}
