% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-analysis.R
\name{get_analysis}
\alias{get_analysis}
\title{Get analysis of the environment for debugging purposes.}
\usage{
get_analysis()
}
\value{
A list with various information regarding the datasets.
}
\description{
Get analysis of the environment for debugging purposes.
}
