% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-annotation-info.R
\name{get_annotation_info}
\alias{get_annotation_info}
\title{Get the annotation_info from the dataset.}
\usage{
get_annotation_info(dataset)
}
\arguments{
\item{dataset}{the dataset object}
}
\value{
the annotation_info tibble
}
\description{
Get the annotation_info from the dataset.
}
