% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-annotations.R
\name{get_annotations}
\alias{get_annotations}
\title{Get the annotations from the dataset.}
\usage{
get_annotations(dataset)
}
\arguments{
\item{dataset}{the dataset object}
}
\value{
the annotations tibble
}
\description{
Get the annotations from the dataset.
}
