% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-correlation.R
\name{get_correlation_plot_data}
\alias{get_correlation_plot_data}
\title{Get the correlation data for plotting.}
\usage{
get_correlation_plot_data(
  dataset,
  id,
  dataset_correlate,
  id_correlate,
  intcovar = NULL,
  use_qr = TRUE,
  backfill_NA = TRUE
)
}
\arguments{
\item{dataset}{The dataset object}

\item{id}{The unique id in the dataset.}

\item{dataset_correlate}{The dataset to correlate against, dataset if INVALID
or NULL.}

\item{id_correlate}{The identifier from the correlate dataset.}

\item{intcovar}{The interactive covariate.}

\item{use_qr}{\code{TRUE} to use qr decomposition.}

\item{backfill_NA}{\code{TRUE} if we need to impute, we backfill the NAs.}
}
\value{
A named \code{list} with keys of \code{data} and \code{datatypes}.
}
\description{
Perform a correlation analysis on an \code{id} for a \code{dataset}.  By default, the
correlation will be against the same dataset, but a different dataset can
be specified with \code{dataset_correlate}.  Return x and y values for plotting.
}
