% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-covar-matrix.R
\name{get_covar_matrix}
\alias{get_covar_matrix}
\title{Create the covar.matrix element.}
\usage{
get_covar_matrix(dataset, id = NULL)
}
\arguments{
\item{dataset}{the dataset object (synchronized)}

\item{id}{the phenotype identifier, for phenotype datasets}
}
\value{
a named list list with 2 elements, covar_matrix and covar_formula
}
\description{
Create the covar.matrix element.
}
