% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_data}
\alias{get_data}
\title{Get the data from a data from the dataset.}
\usage{
get_data(ds, data_name = NULL)
}
\arguments{
\item{ds}{a dataset object (synchronized or not)}

\item{data_name}{A string containing which data element from the dataset's
data element}
}
\value{
the data element
}
\description{
The data element in the dataset list should either be a matrix or a named
list with each eleemnt being a matrix.
}
\details{
If \code{data_name} is not specified, the order that will be returned is:
data (if a matrix)
rz, norm, raw, log, transformed (if a named list).
}
