% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_dataset_info}
\alias{get_dataset_info}
\title{Get all "dataset.*" information}
\usage{
get_dataset_info()
}
\value{
A named list of all the dataset objects, along with the
  ensembl.version.
}
\description{
This will return a named list of all datasets and the ensmebl version.
}
