% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_dataset_stats}
\alias{get_dataset_stats}
\title{Get all "dataset.*" statistics.}
\usage{
get_dataset_stats()
}
\value{
A named list of all the dataset objects.
}
\description{
This will return a named list of all datasets and some statistics.
}
