% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-expression.R
\name{get_expression}
\alias{get_expression}
\title{Get the expression data.}
\usage{
get_expression(dataset, id)
}
\arguments{
\item{dataset}{the dataset object}

\item{id}{the unique id in the dataset}
}
\value{
a named `list` with 2 elements.
data - a `tibble` with samples and data
datatypes - a `list` of all the types and the unique values
}
\description{
Get the expression data.
}
