% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-lod-peaks.R
\name{get_lod_peaks}
\alias{get_lod_peaks}
\title{Get the LOD peaks for a dataset.}
\usage{
get_lod_peaks(ds)
}
\arguments{
\item{ds}{The dataset object to get the LOD peaks for.}
}
\value{
A named `list` of all the LOD peaks. Each element in the `list` is a
`tibble` with the following columns depending upon `dataset$datatype`:
\itemize{
  \item mRNA - `marker,chr,pos,gene_id,symbol,gene_chrom,middle,lod`
  \item protein - `marker,chr,pos,protein_id,gene_id,symbol,gene_chrom,middle,lod`
  \item phenotype - `marker,chr,pos,data_name,short_name,description,lod`
}

If the allele effects are stored, values `A-H` will also be included.
}
\description{
By default, this will get all LOD peaks stored in the `dataset$lod_peaks`
`list`.  If `intcovar` is specified, only the peaks for that are returned.
}
