% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-lod-peaks.R
\name{get_lod_peaks_all}
\alias{get_lod_peaks_all}
\title{Get the LOD peaks for additive and all covariates.}
\usage{
get_lod_peaks_all(ds)
}
\arguments{
\item{ds}{the dataset object}
}
\value{
a data.frame with the following columns: marker, chr, pos
and depending upons dataset$dataType the following columns:
mRNA = gene_id, symbol, gene_chrom, middle, lod
protein = protein_id, gene_id, symbol, gene_chrom, middle, lod
phenotype = data_name, short_name, description, lod
}
\description{
Get the LOD peaks for additive and all covariates.
}
