% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-lod-peaks-by-covar.R
\name{get_lod_peaks_by_covar}
\alias{get_lod_peaks_by_covar}
\title{Get the LOD peaks for a dataset.}
\usage{
get_lod_peaks_by_covar(ds, intcovar = NULL, full_annotations = FALSE)
}
\arguments{
\item{ds}{The dataset object to get the LOD peaks for.}

\item{intcovar}{NULL for 'additive' or a value representing the interactive
covariate.}

\item{full_annotations}{`TRUE` to send back full annotations.}
}
\value{
A a `tibble` with the following columns depending upon
`dataset$datatype`.

If the allele effects are stored, values `A-H` will also be included.
}
\description{
Used internally to the package.
}
\details{
By default, this will get all LOD peaks stored in the `dataset$lod_peaks`
`list`.  If `intcovar` is specified, only the peaks for that are returned.
}
