% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-lod-scan.R
\name{get_lod_scan_by_sample}
\alias{get_lod_scan_by_sample}
\title{Perform the LOD scan for each "value" of the interactive covariate.}
\usage{
get_lod_scan_by_sample(ds, id, intcovar, chrom, cores = 0)
}
\arguments{
\item{ds}{the dataset object}

\item{id}{the unique id in the dataset}

\item{intcovar}{the interactive covariate}

\item{chrom}{The chromosome.}

\item{cores}{number of cores to use (0=ALL)}
}
\value{
A named list with each name a "sample value" and the element is a
  tibble with the following columns: id, chr, pos, lod.
}
\description{
Perform the LOD scan for each "value" of the interactive covariate.
}
