% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-markers.R
\name{get_markers}
\alias{get_markers}
\title{Get the markers (more for API benefits).  Not for qtl2.}
\usage{
get_markers(chrom = NULL)
}
\arguments{
\item{chrom}{The chromosome to filter on.}
}
\value{
A tibble with the following columns: marker_id, chr, bp
}
\description{
Get the markers (more for API benefits).  Not for qtl2.
}
