% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-mediation.R
\name{get_mediation}
\alias{get_mediation}
\title{Perform mediation}
\usage{
get_mediation(dataset, id, marker_id, dataset_mediate = NULL)
}
\arguments{
\item{dataset}{the dataset object}

\item{id}{the unique id in the dataset}

\item{marker_id}{marker identifier}

\item{dataset_mediate}{the dataset object to mediate against}
}
\value{
A named list with the covar_formula and data that is a tibble with
        the following columns depending on datatype:
        mRNA = gene_id, symbol, chr, pos, LOD
        protein = protein_id, gene_id, symbol, chr, pos, LOD
        phenotype = NONE
}
\description{
Perform mediation
}
