% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_random_id}
\alias{get_random_id}
\title{Get a random identifier from the dataset.}
\usage{
get_random_id(dataset)
}
\arguments{
\item{dataset}{a dataset object}
}
\value{
a random identifier
}
\description{
Get a random identifier from the dataset.
}
