% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-rankings.R
\name{get_rankings}
\alias{get_rankings}
\title{Calculate the rankings for which gene/protein should be shown.}
\usage{
get_rankings(dataset, chrom = NULL, min_value = 100, max_value = 1000)
}
\arguments{
\item{dataset}{The dataset object.}

\item{chrom}{The chromosome to filter on.}

\item{min_value}{Minimum score value (defaults to 100).}

\item{max_value}{Maximum score value (defaults to 1000).}
}
\value{
A \code{tibble} with the identifiers and rankings.
}
\description{
Calculate the rankings for which gene/protein should be shown.
}
