% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-samples.R
\name{get_sample_id_field}
\alias{get_sample_id_field}
\title{Get the sample id field.}
\usage{
get_sample_id_field(ds)
}
\arguments{
\item{ds}{the dataset object}
}
\value{
the name of the sample id field
}
\description{
Get the sample id field.
}
