% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-samples.R
\name{get_samples}
\alias{get_samples}
\title{Get the samples from the dataset.}
\usage{
get_samples(dataset)
}
\arguments{
\item{dataset}{the dataset object}
}
\value{
the samples tibble
}
\description{
Get the samples from the dataset.
}
