% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-snp-assoc-mapping.R
\name{get_snp_assoc_mapping}
\alias{get_snp_assoc_mapping}
\title{Get the SNP association mapping}
\usage{
get_snp_assoc_mapping(
  ds,
  id,
  intcovar,
  chrom,
  location,
  db_file,
  window_size = 5e+05,
  cores = 0
)
}
\arguments{
\item{ds}{the dataset object}

\item{id}{the unique id in the dataset}

\item{intcovar}{the interactive covariate}

\item{chrom}{The chromosome.}

\item{location}{location on chromosome in base pairs}

\item{db_file}{full path to the sqlite database file}

\item{window_size}{the size of the window to scan before and after location}

\item{cores}{number of cores to use (0=ALL)}
}
\value{
a `data.frame` with the following columns: snp, chr, pos, alleles,
  sdp, ensembl_gene, csq, index, interval, on_map, lod.
}
\description{
Get the SNP association mapping
}
