% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{id_exists}
\alias{id_exists}
\title{Check if id exists and has data in dataset.}
\usage{
id_exists(id)
}
\arguments{
\item{id}{the id to check}
}
\value{
\code{TRUE} if id contains data, \code{FALSE} otherwise
}
\description{
Check if id exists and has data in dataset.
}
