% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{invalid}
\alias{invalid}
\title{Test if `val` is invalid.}
\usage{
invalid(val)
}
\arguments{
\item{val}{value to be tested}
}
\value{
TRUE if invalid, FALSE otherwise
}
\description{
Test if `val` is invalid.
}
