% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_phenotype}
\alias{is_phenotype}
\title{Check dataset to see if the datatype value is "phenotype"}
\usage{
is_phenotype(ds)
}
\arguments{
\item{ds}{a dataset object}
}
\value{
`TRUE` if the datatype is phenotype, `FALSE` otherwise
}
\description{
Check dataset to see if the datatype value is "phenotype"
}
