% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nvl}
\alias{nvl}
\title{Check value for validity and return it or a default}
\usage{
nvl(value, default)
}
\arguments{
\item{value}{value to check}

\item{default}{default value to use if value is "invalid"}
}
\value{
`value` if it is valid, `default` otherwise
}
\description{
Check value for validity and return it or a default
}
