% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nvl_int}
\alias{nvl_int}
\title{Convert value to numeric if possible}
\usage{
nvl_int(value, default)
}
\arguments{
\item{value}{value to check}

\item{default}{default value to use if value is "invalid"}
}
\value{
\code{value} if it is numeric, \code{default} otherwise
}
\description{
Convert value to numeric if possible
}
