% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{synchronize_data}
\alias{synchronize_data}
\title{Synchronize data and subset accordingly.}
\usage{
synchronize_data(dataset)
}
\arguments{
\item{dataset}{the dataset object}
}
\value{
list with 3 elements: `annots`, `samples` and `data`.
}
\description{
Synchronize data and subset accordingly.
}
