% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{synchronize_dataset}
\alias{synchronize_dataset}
\title{Get the dataset that is altered to be better use in the environment.}
\usage{
synchronize_dataset(dataset)
}
\arguments{
\item{dataset}{the dataset element}
}
\value{
the altered dataset element
}
\description{
Get the dataset that is altered to be better use in the environment.
}
