% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-samples.R
\name{synchronize_samples}
\alias{synchronize_samples}
\title{Synchronize sample IDs between objects.}
\usage{
synchronize_samples(pheno, probs, expr, covar = NULL)
}
\arguments{
\item{pheno}{a matrix or data.frame containing the phenotypes}

\item{probs}{a numeric matrix containing the founder allele probabilities}

\item{expr}{a numeric matrix containing the gene expression data}

\item{covar}{a numeric matrix containing the mapping covariates}
}
\value{
list with four elements: `pheno`, `probs`, `expr` and `covar`. The
Sample IDs will all be in the same order.
}
\description{
If each object exists, we get the intersection of the sample IDs,
sort them and subset each object.  Sample IDs must be in rownames for
each parameter.
}
