% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valid}
\alias{valid}
\title{Test if \code{val} is valid.}
\usage{
valid(val)
}
\arguments{
\item{x}{value to be tested}
}
\value{
TRUE if valid, FALSE otherwise
}
\description{
Test if \code{val} is valid.
}
