% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_annotations}
\alias{validate_annotations}
\title{Validate the annotations in the dataset to make sure qtl2api can use it.}
\usage{
validate_annotations(dataset)
}
\arguments{
\item{dataset}{the dataset as a string identifier}
}
\description{
Validate the annotations in the dataset to make sure qtl2api can use it.
}
