% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_data}
\alias{validate_data}
\title{Validate the data in the dataset to make sure qtl2api can use it.}
\usage{
validate_data(dataset)
}
\arguments{
\item{dataset}{the dataset_id as a string identifier}
}
\description{
Validate the data in the dataset to make sure qtl2api can use it.
}
