% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_environment}
\alias{validate_environment}
\title{Validate the environment for defined objects.}
\usage{
validate_environment(extensive = FALSE)
}
\arguments{
\item{extensive}{TRUE to perform scans on random elements in the data}
}
\description{
Validate the environment for defined objects.
}
