% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_lod_peaks}
\alias{validate_lod_peaks}
\title{Validate the lod.peaks to make sure qtl2api can use it.}
\usage{
validate_lod_peaks(dataset)
}
\arguments{
\item{dataset}{the dataset as a string identifier}
}
\description{
Validate the lod.peaks to make sure qtl2api can use it.
}
