% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_samples}
\alias{validate_samples}
\title{Validate the annot.samples in the dataset to make sure qtl2api can use it.}
\usage{
validate_samples(dataset_id)
}
\arguments{
\item{dataset_id}{the dataset_id as a string identifier}
}
\description{
Validate the annot.samples in the dataset to make sure qtl2api can use it.
}
