% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-lod-scores.R
\name{calc_lod_scores}
\alias{calc_lod_scores}
\title{Perform a LOD scan.}
\usage{
calc_lod_scores(
  dataset,
  id,
  intcovar = NULL,
  cores = 0,
  filter_threshold = 6,
  filter_peak_drop = Inf,
  filter_thresholdX = NULL,
  filter_peak_dropX = NULL,
  scan1_output = FALSE,
  cis_window = 1e+07
)
}
\arguments{
\item{dataset}{The dataset object.}

\item{id}{The unique id in the dataset.}

\item{intcovar}{The interactive covariate.}

\item{cores}{Number of cores to use (0 = ALL).}

\item{filter_threshold}{If set, qtl2::find_peaks is used.}

\item{filter_peak_drop}{If set, qtl2::find_peaks is used.}

\item{filter_thresholdX}{If set, qtl2::find_peaks is used.}

\item{filter_peak_dropX}{If set, qtl2::find_peaks is used.}

\item{scan1_output}{If `TRUE`, original `qtl2::scan1` data is included.}

\item{cis_window}{cis defined window}
}
\value{
a `list` with the following elements:
\itemize{
  \item lod_peaks - `tibble` of LOD peaks
  \item lod_scores - `tibble` with the following columns: id, chr, pos, lod
  \item scan1 - `qtl2::scan1` output
}
}
\description{
Perform a LOD scan.
}
