% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-correlation.R
\name{get_correlation}
\alias{get_correlation}
\title{Get the correlation.}
\usage{
get_correlation(dataset, id, dataset_correlate = NULL, intcovar = NULL)
}
\arguments{
\item{dataset}{the dataset object}

\item{id}{the unique id in the dataset}

\item{dataset_correlate}{the dataset to correlate against, dataset if INVALID}

\item{intcovar}{the interactive covariate}
}
\value{
a `tibble` with the correlation and annotations
}
\description{
Get the correlation.
}
