% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-correlation.R
\name{get_correlation_plot_data}
\alias{get_correlation_plot_data}
\title{Get the correlation data for plotting.}
\usage{
get_correlation_plot_data(
  dataset,
  id,
  dataset_correlate,
  id_correlate,
  intcovar = NULL,
  use_qr = TRUE,
  backfill_NA = TRUE
)
}
\arguments{
\item{dataset}{the dataset object}

\item{id}{the unique id in the dataset}

\item{dataset_correlate}{the dataset to correlate to}

\item{id_correlate}{the identifier from the correlate dataset}

\item{intcovar}{the interactive covariate}

\item{use_qr}{qr decomposition}

\item{backfill_NA}{if we need to impute, should we backfill the NAs}
}
\value{
a named `list` with the data to plot
}
\description{
Get the correlation data for plotting.
}
