% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_covar_matrix}
\alias{get_covar_matrix}
\title{Use the existing covar.matrix or create it.}
\usage{
get_covar_matrix(ds, id = NULL)
}
\arguments{
\item{ds}{the dataset object}

\item{id}{the phenotype identifier, for phenotype datasets}
}
\value{
the covar matrix
}
\description{
Use the existing covar.matrix or create it.
}
