% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_covar_matrix}
\alias{get_covar_matrix}
\title{Creaste the covar.matrix element.}
\usage{
get_covar_matrix(dataset, id = NULL)
}
\arguments{
\item{dataset}{the dataset object}

\item{id}{the phenotype identifier, for phenotype datasets}
}
\value{
the covar matrix
}
\description{
Creaste the covar.matrix element.
}
