% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_dataset_by_id}
\alias{get_dataset_by_id}
\title{Get the dataset by id (a string).  annot.samples can be used at the top level
to share sample annotations among other datasets.}
\usage{
get_dataset_by_id(dataset_id)
}
\arguments{
\item{dataset_id}{a string, either 'dataset.name' or just 'name'}
}
\value{
the dataset element
}
\description{
Get the dataset by id (a string).  annot.samples can be used at the top level
to share sample annotations among other datasets.
}
