% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-expression.R
\name{get_expression}
\alias{get_expression}
\title{Get the expression data.}
\usage{
get_expression(dataset, id)
}
\arguments{
\item{dataset}{The dataset object.}

\item{id}{The unique id in the dataset.}
}
\value{
a named \code{list} with 2 elements.
\itemize{
\item data - a \code{tibble} with samples and data
\item datatypes - a \code{list} of all the types and the unique values
}
}
\description{
Get the expression data.
}
