% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-expression.R
\name{get_expression}
\alias{get_expression}
\title{Get the expression data.}
\usage{
get_expression(dataset, id)
}
\arguments{
\item{dataset}{the dataset object}

\item{id}{the unique id in the dataset}
}
\value{
a `list` with elements of tibble and list of the datatypes.
}
\description{
Get the expression data.
}
