% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-founder-coefficients.R
\name{get_founder_coefficients}
\alias{get_founder_coefficients}
\title{Get the founder coefficients}
\usage{
get_founder_coefficients(
  dataset,
  id,
  chrom,
  intcovar = NULL,
  blup = FALSE,
  center = TRUE,
  cores = 0
)
}
\arguments{
\item{dataset}{the dataset object}

\item{id}{the unique id in the dataset}

\item{chrom}{the chromosome}

\item{intcovar}{the interactive covariate}

\item{blup}{whether or not to perform BLUP}

\item{center}{whether or not to center the data}

\item{cores}{number of cores to use (0=ALL)}
}
\value{
a named `list` with each element being a tibble with the following
        columns: id, chr, pos, and A-H, also the covar_formula used
}
\description{
Get the founder coefficients
}
