% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-lod-peaks.R
\name{get_lod_peaks}
\alias{get_lod_peaks}
\title{Get the LOD peaks}
\usage{
get_lod_peaks(ds, intcovar = NULL)
}
\arguments{
\item{ds}{the dataset object}

\item{intcovar}{the interactive covariate}
}
\value{
a `data.frame` with the following columns: marker, chr, bp
and depending upon dataset$dataType the following columns:
mRNA = gene_id, symbol, gene_chrom, middle, lod
protein = protein_id, gene_id, symbol, gene_chrom, middle, lod
phenotype = data_name, short_name, description, lod
}
\description{
Get the LOD peaks
}
