% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-mediation.R
\name{get_mediation}
\alias{get_mediation}
\title{Perform mediation}
\usage{
get_mediation(ds, id, marker_id, ds_mediate = NULL)
}
\arguments{
\item{ds}{the dataset object}

\item{id}{the unique id in the dataset}

\item{marker_id}{marker identifier}

\item{ds_mediate}{the dataset object to mediate against}
}
\value{
A data.frame with the following columns depending on datatype:
        mRNA = gene_id, symbol, chr, pos, LOD
        protein = protein_id, gene_id, symbol, chr, pos, LOD
        phenotype = NONE
}
\description{
Perform mediation
}
