% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-mediation.R
\name{get_mediation}
\alias{get_mediation}
\title{Perform mediation.}
\usage{
get_mediation(dataset, id, marker_id, dataset_mediate = NULL)
}
\arguments{
\item{dataset}{The dataset object.}

\item{id}{The unique id in the dataset.}

\item{marker_id}{The unique marker identifier.}

\item{dataset_mediate}{The dataset object to mediate against (defaults to
dataset)}
}
\value{
A \code{tibble} with the following columns depending on datatype:
\itemize{
\item mRNA - gene_id, symbol, chr, pos, LOD
\item protein = protein_id, gene_id, symbol, chr, pos, LOD
\item phenotype = NONE
}
}
\description{
Perform mediation.
}
