% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-rankings.R
\name{get_rankings}
\alias{get_rankings}
\title{Calculate the rankings for which gene/protein should be shown.}
\usage{
get_rankings(dataset, chrom = NULL, min_value = 100, max_value = 1000)
}
\arguments{
\item{dataset}{the dataset object}

\item{chrom}{the chromosome to filter on}

\item{min_value}{minimum score value (defaults to 100).}

\item{max_value}{maximum score value (defaults to 1000).}
}
\value{
A tibble with the following columns: id, ranking
}
\description{
Calculate the rankings for which gene/protein should be shown.
}
