% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{id_exists}
\alias{id_exists}
\title{Check if id exists and has data in dataset.}
\usage{
id_exists(id, ds = NULL)
}
\arguments{
\item{id}{the id to check}

\item{ds}{a dataset object}
}
\value{
`TRUE` if id contains data, `FALSE` otherwise
}
\description{
Check if id exists and has data in dataset.
}
