% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_covar_info}
\alias{validate_covar_info}
\title{Validate the covar.info in the dataset to make sure qtl2api can use it.}
\usage{
validate_covar_info(dataset_id)
}
\arguments{
\item{dataset_id}{the dataset_id as a string identifier}
}
\description{
Validate the covar.info in the dataset to make sure qtl2api can use it.
}
