% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_dataset}
\alias{validate_dataset}
\title{Validate the dataset to make sure qtl2api can use it.}
\usage{
validate_dataset(dataset, extensive = FALSE)
}
\arguments{
\item{dataset}{the dataset as a string identifier}

\item{extensive}{TRUE to perform scans on random elements in the data}
}
\description{
Validate the dataset to make sure qtl2api can use it.
}
