% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_samples}
\alias{validate_samples}
\title{Validate the annot_samples in the dataset to make sure qtl2api can use it.}
\usage{
validate_samples(dataset)
}
\arguments{
\item{dataset}{the dataset as a string identifier}
}
\description{
Validate the annot_samples in the dataset to make sure qtl2api can use it.
}
